pyproject.toml
[tool.poetry]
name = "chess-openings-ai"
version = "0.1.0"
description = ""
authors = ["Travis John Villanueva <vtravisjohn@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.38.0"
langchain = "^0.3.0"
google-cloud-aiplatform = "^1.67.1"
chromadb = "^0.5.7"
python-dotenv = "^1.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
